Created the preferences window
Its not completed but its there. I also deleted the ErrorBuffer, AFAIK
we don't need it. The switch statement that controls the error messages
was cleaned up. The radiobutton on the preferences window is not showing
up, I will probally figure it out soon. Also added a few comments here
and there.