Upgraded download and extract engine
I spent quite a bit of time upgrading the extract engine. It now return all the formats, in a vector (including the real_urls)
It has a few validation checks before it tried to grab the format and real_url.
I've also added a wxMsg to current_videos.extension in MainFrame.cpp, incase you have to login to view the video, or that it is blocked by region.
My next push should contain comments, as I have not put them in yet.